+++ packages/go_router/lib/src/delegate.dart
@@ -133,6 +133,7 @@ class GoRouterDelegate extends RouterDelegate<RouteMatchList>
       return false;
     }
     _matchList.pop();
+    notifyListeners();
     assert(() {
       _debugAssertMatchListNotEmpty();
       return true;
@@ -144,7 +145,7 @@ class GoRouterDelegate extends RouterDelegate<RouteMatchList>
   ///
   /// See also:
   /// * [push] which pushes the given location onto the page stack.
-  void replace(RouteMatchList matches) {
+  void pushReplacement(RouteMatchList matches) {
     _matchList.pop();
     push(matches); // [push] will notify the listeners.
   }